Prefer current song over playlist entry for notification
The contents of status.playlist.nowPlaying and the playlist entry
pointed to by status.playlist.currentPosition can differ, e.g. in case
of radio streams that regularly update their title to reflect the song
being played. In that case, the playlist timestamp isn't updated, which
means latestPlaylist isn't updated either, which means that the playlist
entry is stuck at the contents of the latest playlist change.
Handle that situation by always prefering status.playlist.nowPlaying in
that case, which is updated properly.